+++ src/main/java/com/example/io_backend/repository/AdditionalServicesRepository.java
@@ -3,5 +3,5 @@ package com.example.io_backend.repository;
 import com.example.io_backend.model.AdditionalServices;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdditionalServicesRepository extends JpaRepository<AdditionalServices, Long> {
+public interface AdditionalServicesRepository extends JpaRepository<AdditionalServices, Integer> {
 }
