+++ src/main/java/com/example/io_backend/repository/LocationRepository.java
@@ -3,5 +3,5 @@ package com.example.io_backend.repository;
 import com.example.io_backend.model.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationRepository extends JpaRepository<Location, Long> {
+public interface LocationRepository extends JpaRepository<Location, Integer> {
 }
