+++ src/main/java/com/example/io_backend/repository/VictimRepository.java
@@ -3,5 +3,5 @@ package com.example.io_backend.repository;
 import com.example.io_backend.model.Victim;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VictimRepository extends JpaRepository<Victim, Long> {
+public interface VictimRepository extends JpaRepository<Victim, Integer> {
 }
